refactor: consolidate database and queue operations into deps package
- Moved MongoDB client creation from database package to deps
- Moved RabbitMQ operations from queue package to deps
- Moved events consumer from events package to deps
- Updated main.go to use consolidated deps package functions
- Removed redundant packages (database, queue, events)
