+++ pkg/deps/consumer.go
@@ -1,11 +1,10 @@
-package events
+package deps
 
 import (
 	"context"
 
 	amqp "github.com/rabbitmq/amqp091-go"
 	"github.com/vit0rr/publish-and-consume/config"
-	"github.com/vit0rr/publish-and-consume/pkg/deps"
 	"github.com/vit0rr/publish-and-consume/pkg/log"
 	"github.com/vit0rr/publish-and-consume/shared"
 	"go.mongodb.org/mongo-driver/mongo"
@@ -17,12 +16,12 @@ type Message struct {
 }
 
 type Consumer struct {
-	deps     *deps.Deps
+	deps     *Deps
 	RabbitMQ *amqp.Connection
 	Mongo    *mongo.Database
 }
 
-func NewConsumer(deps *deps.Deps, amqpConn *amqp.Connection, db *mongo.Database, config config.Config, ctx context.Context) (*Consumer, error) {
+func NewConsumer(deps *Deps, amqpConn *amqp.Connection, db *mongo.Database, config config.Config, ctx context.Context) (*Consumer, error) {
 	return &Consumer{
 		deps:     deps,
 		RabbitMQ: amqpConn,
